Convert Daily data of a stock to Monthly data

Convert Daily data of a stock to Monthly data

daytomonth(x)

Arguments

x

dataframe containing stock data with columns SYMBOL, OPEN, HIGH, LOW, CLOSE & DATE.

Value

OHLC data in monthly format.

Details

Converts the Daily data of a stock to Monthly data.

Note

The dataframe should contain six columns named SYMBOL, OPEN, HIGH, LOW, CLOSE & DATE.

See also

Author

Nandan L. Patil tryanother609@gmail.com

Examples

# \donttest{
data(dailydata)
daytomonth(dailydata)
#>    SYMBOL   OPEN   HIGH    LOW  CLOSE          MONTH
#> 1    SBIN 250.10 317.80 244.35 312.40   October-2019
#> 2    SBIN 312.40 351.00 299.70 341.85  November-2019
#> 3    SBIN 343.90 344.60 308.00 333.75  December-2019
#> 4    SBIN 334.70 339.85 305.65 318.45   January-2020
#> 5    SBIN 317.95 331.90 295.35 303.00  February-2020
#> 6    SBIN 311.00 312.00 173.55 196.85     March-2020
#> 7    SBIN 194.00 198.00 175.00 190.50     April-2020
#> 8    SBIN 182.45 183.50 149.45 161.30       May-2020
#> 9    SBIN 164.00 197.50 163.35 178.45      June-2020
#> 10   SBIN 179.45 202.50 178.60 191.45      July-2020
#> 11   SBIN 192.00 231.55 189.55 212.00    August-2020
#> 12   SBIN 213.10 219.00 175.50 185.40 September-2020
#> 13   SBIN 187.00 207.30 185.90 189.25   October-2020
#> 14   SBIN 192.20 253.00 190.05 244.25  November-2020
#> 15   SBIN 245.00 279.90 244.10 274.95  December-2020
#> 16   SBIN 274.90 310.90 269.50 282.10   January-2021
#> 17   SBIN 285.10 427.70 282.75 390.15  February-2021
#> 18   SBIN 395.10 408.90 345.20 364.30     March-2021
#> 19   SBIN 367.70 371.90 321.30 353.50     April-2021
#> 20   SBIN 349.60 433.65 341.40 424.35       May-2021
#> 21   SBIN 426.05 441.95 400.50 419.20      June-2021
#> 22   SBIN 420.30 444.40 417.15 431.80      July-2021
#> 23   SBIN 434.75 467.45 401.25 426.05    August-2021
#> 24   SBIN 427.50 471.90 425.10 453.00 September-2021
#> 25   SBIN 448.50 471.00 443.80 457.90   October-2021
# }